Add Web3Implementation as a superclass for Ethereum and xDai blockchain implementations
