+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules
 .DS_Store
+desktop.ini
 .idea
 .vscode/*
 .env
